refactor: move plugin module to its own file
